Adds `break` and `continue` tags.
Adds the notion of an "interrupt" - a signal in the rendering context
that stops all rendering and returns the rendered content back up the
call stack. When the stack unwind hits a for_loop, it resets the
interrupt and continues processing as directed - either breaking out
of the loop or continuing on to the next iteration.