Docs, remove ErrorMode and FilterResult
This commit adds some documentation to the Liquid API.

It also removes the ErrorMode. ErrorMode is a (bad) leftover from the
original Liquid for Ruby, however it is very much about shared global
state and I can't see us supporting it in the Rust version. It was a bad
idea anyway.

Also removed was FilterResult, which was simply an alias for
Result<FilterError>.