Allow for variables to be passed as filter args
This commit adds support for passing variables as arguments to filter.
It changes the FilterPrototype type to contain a list of VarOrVal
instances rather than just Value instances.

When filters are being applied, each VarOrVal instance is "unwrapped"
before being given to the filter invocation. For VarOrVal::Val this
means just extracting the inner Value. For VarOrVal::Var it means
looking up the variable value.

This fixes #49