feat: Implement `contains` operator
The contains operator works for strings (substring search), arrays
(string in array) or objects (string key exists).
The renderer will error if the right-hand side operator does not
evaluate to a string, if the left-hand side is neither a string, array
nor object or if the array contains non-string items.

Fixes #155