fix(value): Improve value coercion practices
Template impact: There is now a consistent approach taken to when a
`Value` can be used as a string, whole number, fractional number, bool
etc.

Examples
- Every `Value` is now treated as a string.
- Whole numbers are now preserved.
- Whole numbers can convert to fractional numbers but not the other way
  around.
- You can convert a fractional number to a whole number using the
  filters `round`, `ceil`, or `floor`.
- Operations on only whole numbers preserve their whole number nature.
- for tag ranges, indexing, etc only accept whole numbers

Fixes #99

Programmatic impact:  This implements the `Value` refactor from #95.
`Value` is still an enum but now there is a single `Scalar` variant.
This scalar is not an enum but is convertible between each of the
common data types.  Optimizations are in place to natively store some of
these data types.

BREAKING CHANGE: `Value`'s API has changed. Conversion behavior within
templates has changed.