feat(date): Support today/now
I would like to find a way to make `now` only capture the value once but
not seeing a good way to do that.  If this was being processed at a
higher level, then we could have the user set it in the `Context`.  A
`lazy_static` wouldn't work because the process could be long running,
serving multiple templates.

Fixes #181