fix: Deeply nested array indexes
The parser is pretty messed up.  I have no idea why it does what it does
but it has two different code paths for parsing variables.  In one, it
somehow was treationg `a.b` as a variable `"a.b"` and not `a["b"]`.

Not sure why it broke but it now works.

Fixes #230