feat(interpreter): Allow named stack frames
The goal of this is so that runtime includes of partials can set this
and then any error (usually the `trace`) that happens within that partial can be prefixed with
the name.  Eventually we also want these to include line numbers (see #247).

In the end, this will make the errors look more like what you'd expect
from a compiler error.