fix(filters): Sort filters now handle nil
Sort filters now handle nil correctly: nil is bigger than everything
else. This is implemented closely following Ruby's implementation:
https://github.com/Shopify/liquid/blob/
8dfe198e168eef580345842ff49c4f2549e238b/lib/liquid/
standardfilters.rb#L441
Since this is implemented explicitly for the sort filter, I chose not
to implement this as a part of the native comparison between `Value`s.
A slight variation is that `nil_safe_compare(nil, nil)` is 0 and not
`1`.
This fixes issue 262.