fix(filters): Sort filters support property
The `sort` and `sort_natural` support an optional property parameter
used to sort objects. If a property is given the filter checks that all
the elements in the array are objects.
If the property is not present in an element in the array, its "sorting
key" is considered to be `nil`, and is treated as it would be in a
plain invocation of `sort{,_natural}`. If more than one element does
not contain the attibute, the order between them is undefined. In order
to address that, the tests now check that one of two possible results
are generated by the filter (see
`test_sort_when_property_is_sometimes_missing_puts_nils_last` and
`test_sort_natural_when_property_is_sometimes_missing_puts_nils_last`).
This solves issues #249, #333, #334. It also solves some of issue #257,
but the filter doesn't yet work on a single hash, as required by the
`test_legacy_sort_hash` test.