fix(filters): Filters work with single values
Introduced a macro called as_sequence, which transforms a Value to
either an iterator, and invokes a code block on it. This tries to match
the Ruby implementation InputIterator
(https://github.com/Shopify/liquid/blob/
98dfe198e168eef580345842ff49c4f2549e238b/lib/liquid/
standardfilters.rb#L457).
This is implemented as macro to avoid dynamic dispatching while keeping
a resonable syntax.
Currently this is used by `sort`, `sort_natural`, and `where`. The
latter already had this functionality, but with another syntax.
This fixes issues #250 and #257.