fix(reflect): Provide access to reflection trait
Rust doesn't seem to support upcasting at the moment, so we need some
way to make it so we can go from the actual Block, Tag, or Filter trait
to the reflection one.  In this case, I decided to just break away from
using supertraits.

BREAKING CHANGE: Changed from reflection being a supertrait to being
exposed in an accessor method.