fix(derive): Depend on liquid-core
Switching to 2018 edtion is blocked on having a crate to depend on.
Currently, we depend on `liquid::`.  Inside of `liquid`, we use a module
`liquid` for this to work.  With 2018 edition, there isn't a common root
for `liquid` and client crates to both use.

The first solution considered was to conditionally compile derive.  That
was going to get too messy.  Indtead, we create a facade around the
basic crates, a kind of SDK.  Plugin development must use this new
`liquid-core`.  As part of this, the interface to `liquid` was
simplified to what is needed for end-users not creating their own
plugins.

BREAKING CHANGE: A lot.  For plugin authors, see the changes we made.
For non-plugin use, most identifiers will now just be at `liquid::`.