feat(runtime)!: Allow scopes to use native types
Previously, when interacting with the stack, we had to convert
application-native types (`dyn ObjectView`) to liquid-native types
(`Object`).  Now we can use `dyn ObjectView` in more places, removing
the need for conversions.

With this also comes support for higher layers in the stack to have
references to lower layers of the stack.

Both of these should significantly reduce the number of allocations
needed for applications.  For example, `forloop` blocks would have to
duplicate the data they were iterating on, which could be quite large,
like an external json file.  Now, `forloop` can be updated to have a
`ValueCow` to lower layers on the stack and wrap all of this up in a
`dyn ObjectView` rather than `Object`.

To make this work, I needed to make Rust aware of the stack structure,
this meant switchin from a `Vec<Frame>` to frames composed on top of
each other.

For mutable globals, we have to allocate on access now,  That is
hopefully going to be ok since we only allocate what is accessed.   The
biggest risk is if someone does a transform on a large data structure
and assigns that to a value.  The biggest loss is we aren't
automatically reporting other root-level variables the user could have
referenced on error.  The problem is knowing which frame is the top of
the stack for successfully calling down into other frames.  I tried to
maintain the existing `find_frame` approach but that was going to
require clonin the entire root-level mutable-global on each access which
seemed like a no-go.

I'm not quite sure if I got all the precedence setup correctly.  We'll
need to audit against the Ruby implementation.

Since so much was shaking up, and to reduce `Cell` use, I switched
`InterruptState` to be `InterruptRegister`.  One less "special case"
being handled.  Next steps would be finding a way to make `PartialStore`
a register and find out how the abstraction for `IndexFrame` could be
generalized.